3-CodigoAdicional/remover-fundo/requirements.txt
Pillow
rembg
streamlit
